Refactor, begin to declare implicitly declared arrays
